Update to webin-cli to v8.0.0, clean up assembly creation functions
