+++ ena-submission/Dockerfile
@@ -20,7 +20,7 @@ RUN micromamba config set extract_threads 1 \
 ARG MAMBA_DOCKERFILE_ACTIVATE=1
 
 
-ENV WEBIN_CLI_VERSION 7.3.1
+ENV WEBIN_CLI_VERSION 8.0.0
 USER root
 RUN wget -q "https://github.com/enasequence/webin-cli/releases/download/${WEBIN_CLI_VERSION}/webin-cli-${WEBIN_CLI_VERSION}.jar" -O /package/webin-cli.jar
 USER $MAMBA_USER
